ui: post_detail 포스트 영역 ui 수정

+++ lib/src/ui/views/post/post_detail/post_detail_screen.dart
@@ -7,6 +7,7 @@ import 'package:zamongcampus/src/business_logic/utils/methods.dart';
 import 'package:zamongcampus/src/business_logic/view_models/post_detail_screen_viewmodel.dart';
 import 'package:zamongcampus/src/config/service_locator.dart';
 import 'package:zamongcampus/src/config/size_config.dart';
+import 'package:zamongcampus/src/ui/common_components/custom_alert_dialog_components/delete/post_deleted_msg.dart';
 import 'package:zamongcampus/src/ui/common_components/custom_alert_dialog_components/report_post/report_post_form.dart';
 import 'package:zamongcampus/src/ui/common_components/sub_appbar_components/sub_appbar.dart';
 import 'package:zamongcampus/src/ui/common_widgets/isLoading.dart';
@@ -54,9 +55,11 @@ class _PostDetailScreenState extends State<PostDetailScreen> {
                   vm.removeNestedCommentOverlay();
                   Navigator.of(context).pop();
                 },
+                titleText: '게시글',
+                isCenter: true,
                 actions: [
                   IconButton(
-                    icon: const Icon(CupertinoIcons.ellipsis),
+                    icon: const Icon(CupertinoIcons.ellipsis_vertical),
                     color: Colors.black,
                     iconSize: kAppBarIconSizeCP,
                     onPressed: () {
@@ -90,7 +93,14 @@ class _PostDetailScreenState extends State<PostDetailScreen> {
                   ),
                 ),
                 onPressed: () {
-                  vm.deletePost(context, vm.postDetail.id);
+                  Navigator.pop(context);
+                  buildCustomAlertDialog(
+                      context: context,
+                      contentWidget: const PostDeletedMsg(),
+                      btnText: '삭제',
+                      press: () {
+                        vm.deletePost(context, vm.postDetail.id);
+                      });
                 },
               )
             : BottomSheetAction(
